services/planner-agent/requirements.txt
kafka-python==2.0.2
redis==5.0.1
google-generativeai==0.3.2
anthropic==0.7.8
ollama==0.1.7
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
PyGithub==2.1.1
pyyaml==6.0.1
jinja2==3.1.2